dev (#76)
* Сдела профиль, еще добавлена тудушка на вывод

* откат

* сделан захардкоженный профиль

* Сделана кнопка добавления промокодов, сделана логика добавления промокодов без срока действия промокода

* Зачисление баллов по qr-коду.
Пофиксил хардкодинг в l10n.
Немного поменял привилегии.
Добавил либу для корректной миграции Enum в psql.
Исправил другие неточности.

* Сделана основная логика по промокодам

* Update create_promocode.py

* Update promocode.py

* Клавиатура выдачи прав

* menu_kb change

* добавил фильтры на привелегии (не везде)

* Refactor

* Добавил добавление/удаление категорий и товаров + магазин без торговли

* Добавил фильтры привелегий еще в пару мест + рефактор того что хочу

* privileges

* btn-cnacel -> btn_cancel

* Pr resolution

* Отформатировал

* small fixes

* Еще один фикс в дев((((((

* Reformat pep8.
Move ADMIN_CHAT_ID and SUPPORT_CHAT_ID to .env

* Reformat pep8. Added .editorconfig

* feature/edit_events123/belyakov (#58)

* middproceess

* Создание и редактирование мероприятий💪
💪
💪

* fix

* Выдача привелегий эвентов💪💪💪

* НАЧИСЛЕНИЕ БАЛЛОВ!!!!!

* FIx неймингов

* fix2

* почистил l10n

* пофиксл поддержку

* Пофиксил бдшку

* Кодстайл фиксы

* Багфиксы и оптимизации после тестирования

* unused

* Еще пару кодстайл фиксов

* Something changes

* middproceess

* Создание и редактирование мероприятий💪
💪
💪

* fix

* Выдача привелегий эвентов💪💪💪

* НАЧИСЛЕНИЕ БАЛЛОВ!!!!!

* FIx неймингов

* fix2

* почистил l10n

* Пофиксил бдшку

* Кодстайл фиксы

* Багфиксы и оптимизации после тестирования

* unused

* Еще пару кодстайл фиксов

* Something changes

* 6 REBASES!!!

* code style

* Changed locale and chat_actions.

---------


* Preserve migrations in release docker-compose.yaml

* Locale and logging changes

* Ask admins by giving points by the visiting event. And apply when visited.

!!! Needs to do migrations of the database

* Add fields for point to PromocodeActivation.py

!!! Needs to do migrations of the database

* Fixed circular import error

* Fixed error with UserFactory (> 64 symbols)

* time_zone Moscow added

* Пофиксил удаление

* callbackfactory fix

* Фикс начисления баллов и расписания

* опечтка

* dt_tz format

* dt_tz2

* Add spam protect and fix filter error

* Telegram id fix

* timezones fixed

* Add file logger. Fix cache dropping. Add single_message.py middleware.

* Add file logger. Fix cache dropping.

* fixed pkapply (#67)

* fixed pkapply

* Update account.py

drop cache

---------


* Fix chat_actions.py error (use it only for messages)

* Fix chat_actions.py error (use it only for messages) (middleware register)

* Fix l10n

* Fix l10n and code style

* Hotfix/timezones fix (#71)

* Нормальные таймзоны + сортнг эвентов

* ДЭЙТПАРСЕР ИМБА

* Update create.py

* Update create.py

---------


* Fix error with privileges and with chat_actions

* ПРОМКОДЫ КУАРКОДЫ (#70)

* Feature/added time on events (#73)

* Добавил что на эвенты можно начислять + 30 минут до и после

* мини рефактор

* Adding Description to events (#69)

* dev (#68)

* Сдела профиль, еще добавлена тудушка на вывод

* откат

* сделан захардкоженный профиль

* Сделана кнопка добавления промокодов, сделана логика добавления промокодов без срока действия промокода

* Зачисление баллов по qr-коду.
Пофиксил хардкодинг в l10n.
Немного поменял привилегии.
Добавил либу для корректной миграции Enum в psql.
Исправил другие неточности.

* Сделана основная логика по промокодам

* Update create_promocode.py

* Update promocode.py

* Клавиатура выдачи прав

* menu_kb change

* добавил фильтры на привелегии (не везде)

* Refactor

* Добавил добавление/удаление категорий и товаров + магазин без торговли

* Добавил фильтры привелегий еще в пару мест + рефактор того что хочу

* privileges

* btn-cnacel -> btn_cancel

* Pr resolution

* Отформатировал

* small fixes

* Еще один фикс в дев((((((

* Reformat pep8.
Move ADMIN_CHAT_ID and SUPPORT_CHAT_ID to .env

* Reformat pep8. Added .editorconfig

* feature/edit_events123/belyakov (#58)

* middproceess

* Создание и редактирование мероприятий💪
💪
💪

* fix

* Выдача привелегий эвентов💪💪💪

* НАЧИСЛЕНИЕ БАЛЛОВ!!!!!

* FIx неймингов

* fix2

* почистил l10n

* пофиксл поддержку

* Пофиксил бдшку

* Кодстайл фиксы

* Багфиксы и оптимизации после тестирования

* unused

* Еще пару кодстайл фиксов

* Something changes

* middproceess

* Создание и редактирование мероприятий💪
💪
💪

* fix

* Выдача привелегий эвентов💪💪💪

* НАЧИСЛЕНИЕ БАЛЛОВ!!!!!

* FIx неймингов

* fix2

* почистил l10n

* Пофиксил бдшку

* Кодстайл фиксы

* Багфиксы и оптимизации после тестирования

* unused

* Еще пару кодстайл фиксов

* Something changes

* 6 REBASES!!!

* code style

* Changed locale and chat_actions.

---------


* Preserve migrations in release docker-compose.yaml

* Locale and logging changes

* Ask admins by giving points by the visiting event. And apply when visited.

!!! Needs to do migrations of the database

* Add fields for point to PromocodeActivation.py

!!! Needs to do migrations of the database

* Fixed circular import error

* Fixed error with UserFactory (> 64 symbols)

* time_zone Moscow added

* Пофиксил удаление

* callbackfactory fix

* Фикс начисления баллов и расписания

* опечтка

* dt_tz format

* dt_tz2

* Add spam protect and fix filter error

* Telegram id fix

* timezones fixed

* Add file logger. Fix cache dropping. Add single_message.py middleware.

* Add file logger. Fix cache dropping.

* fixed pkapply (#67)

* fixed pkapply

* Update account.py

drop cache

---------


---------


* Adding Event Description

* Fix bug with event points

---------


* Fix indents

* Fix profile buttons

* Add first dialog

* Change datetime formatting to pendulum use

* Add l10n format for aiogram dialog.
Generate diagram of dialogs (in DEBUG).
Some changes in locale.ftl.
Add /support command.

* Fix import loop

* Fix other handlers (open dialog does not skip update).

Fixed /support command.
Refactored admin support handler.

Changed ChatAction middleware to native.

* Some fixes, add db method and changes in structure.

---------

+++ telegram_bot/handlers/admin/edit_events/create.py
@@ -1,11 +1,12 @@
-from datetime import datetime
-
+import dateparser
+import pendulum
 from aiogram import Router, types
 from aiogram.fsm.context import FSMContext
 from fluent.runtime import FluentLocalization
 
 from database import async_session
 from database.methods import create_event, get_user_by_telegram_id
+from env import TelegramKeys
 from filters.main import LocalizedTextFilter
 from keyboards.common import edit_events_kb
 from keyboards.inline import cancel_ikb
@@ -64,11 +65,9 @@ async def ask_for_event_name_h(msg: types.Message, state: FSMContext, l10n: Flue
 @create_router.message(EditEventsActions.CHOOSE_EVENT_START_TIME)
 async def ask_for_event_start_time_h(msg: types.Message, state: FSMContext, l10n: FluentLocalization):
 	state_data = await state.get_data()
-	state_data.update(event_start_time=msg.text.strip())
 
-	try:
-		datetime.strptime(msg.text.strip(), "%d.%m.%Y %H:%M")
-	except ValueError:
+	starts_at = msg.text.strip()
+	if dateparser.parse(starts_at) is None:
 		await msg.bot.delete_message(chat_id=msg.chat.id, message_id=msg.message_id)
 		await msg.answer(l10n.format_value("wrong-datetime"))
 		return
@@ -81,6 +80,7 @@ async def ask_for_event_start_time_h(msg: types.Message, state: FSMContext, l10n
 		message_id=state_data.get("event_message_id")
 	)
 
+	state_data.update(event_start_time=starts_at)
 	await state.set_data(state_data)
 	await state.set_state(EditEventsActions.CHOOSE_EVENT_END_TIME)
 
@@ -88,30 +88,63 @@ async def ask_for_event_start_time_h(msg: types.Message, state: FSMContext, l10n
 @create_router.message(EditEventsActions.CHOOSE_EVENT_END_TIME)
 async def ask_for_event_end_time_h(msg: types.Message, state: FSMContext, l10n: FluentLocalization):
 	state_data = await state.get_data()
-	state_data.update(event_end_time=msg.text.strip())
 
-	try:
-		datetime.strptime(msg.text.strip(), "%d.%m.%Y %H:%M")
-	except ValueError:
+	ends_at = msg.text.strip()
+	if dateparser.parse(ends_at) is None:
 		await msg.bot.delete_message(chat_id=msg.chat.id, message_id=msg.message_id)
 		await msg.answer(l10n.format_value("wrong-datetime"))
 		return
 
+	await msg.delete()
+	await msg.bot.edit_message_text(
+		l10n.format_value("ask-for-event-description"),
+		reply_markup=cancel_ikb(l10n),
+		chat_id=msg.chat.id,
+		message_id=state_data.get("event_message_id")
+	)
+
+	state_data.update(event_end_time=ends_at)
 	await state.set_data(state_data)
+	await state.set_state(EditEventsActions.CHOOSE_EVENT_DESCRIPTION)
+
+
+@create_router.message(EditEventsActions.CHOOSE_EVENT_DESCRIPTION)
+async def ask_for_event_description_h(msg: types.Message, state: FSMContext, l10n: FluentLocalization):
+	description = msg.text.strip()
+	state_data = await state.get_data()
+
+	if len(description) > 1000:
+		await msg.delete()
+		await msg.answer(l10n.format_value("event-description-too-long"))
+		return
+
+	if description != "-":
+		state_data.update(description=description)
 
 	async with async_session() as session:
-		creator = await get_user_by_telegram_id(session, msg.from_user.id)
+		user = await get_user_by_telegram_id(session, msg.from_user.id)
+
+		starts_at = dateparser.parse(state_data.get("event_start_time", ''))
+		if starts_at is not None:
+			starts_at = pendulum.instance(starts_at, TelegramKeys.TZ)
+
+		ends_at = dateparser.parse(state_data.get("event_end_time", ''))
+		if ends_at is not None:
+			ends_at = pendulum.instance(ends_at, TelegramKeys.TZ)
+
 		await create_event(
 			session=session,
 			name=state_data.get("event_name"),
 			visit_points=int(state_data.get("visit_points")),
-			creator_id=creator.id,
-			starts_at=datetime.strptime(state_data.get("event_start_time"), "%d.%m.%Y %H:%M"),
-			ends_at=datetime.strptime(state_data.get("event_end_time"), "%d.%m.%Y %H:%M"),
+			creator_id=user.privileges_id,
+			starts_at=starts_at,
+			ends_at=ends_at,
+			description=state_data.get("description")
 		)
 
-	await msg.delete()
-	await msg.bot.delete_message(chat_id=msg.chat.id, message_id=state_data.get("event_message_id"))
-	await msg.answer(l10n.format_value("event-created"), reply_markup=edit_events_kb(l10n))
-	await state.clear()
-	await state.set_state(EditEventsActions.EDIT_EVENTS)
+		await msg.delete()
+		await msg.bot.delete_message(chat_id=msg.chat.id, message_id=state_data.get("event_message_id"))
+		await msg.answer(l10n.format_value("event-created"), reply_markup=edit_events_kb(l10n))
+
+		await state.clear()
+		await state.set_state(EditEventsActions.EDIT_EVENTS)
